Move serialization code to dedicated module
The changes split out serialization-related code from diff_count.rs and
bitvec.rs into a new serde.rs module. This improves code organization by
separating concerns.
